+++ src/__tests__/load.test.ts
@@ -1,5 +1,5 @@
-import Dbc from '../dbc/dbc';
-import {IncorrectFileExtension} from '../dbc/errors';
+import Dbc from '../dbc/Dbc';
+import {IncorrectFileExtension} from '../dbc/Errors';
 
 // All return values validated using Kvaser CANKing
 test('Load: Invalid File', async () => {
