+++ src/index.ts
@@ -2,8 +2,8 @@
  * Main file
  * Simply exports the main uility class so that it can be referenced when imported via an NPM package
  */
-import Dbc from './dbc/dbc';
-import Can from './can/can';
+import Dbc from './dbc/Dbc';
+import Can from './can/Can';
 
 export default Dbc;
